Update main.js
Instead of passing JSON object directly to a Backbone View, the JSON is passed into a collection view, which passes the data into a Collection (which is configured to a Backbone model) for facilitating future modifications. Each model in this collection is then instantiated as a backbone view, nested inside a collection view.

Events on each 'model' view (individual weather item) has also been re-implemented into the backbone view definition (instead of plain jQuery).